# backend/docs/swagger.yaml
basePath: /
definitions:
...
  httperrors.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  services.AccessToken:
    properties:
      token:
        type: string
    type: object
...
  services.DisplayUser:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      last_seen_at:
        type: string
      metadata:
        items:
          type: integer
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  services.LoginUserDTO:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  services.RegisterUserDTO:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  services.UpdateUserDTO:
    properties:
      avatar_url:
        type: string
      metadata:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a messenger application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Messenger API
  version: "1.0"
paths:
...
  /users/id/{id}:
    get:
      consumes:
      - application/json
      description: Get a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DisplayUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Log in a user.
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Login a user
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DisplayUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the current user's profile.
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DisplayUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /users/me/avatar:
    post:
      consumes:
      - application/json
      description: Upload an avatar image for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DisplayUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload avatar
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user.
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Register a user
      tags:
      - users
  /users/username/{username}:
    get:
      consumes:
      - application/json
      description: Get a user by username.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DisplayUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      summary: Get user by username
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"